Extract plot methods into chartPlotter
